Casting net inputs to float32.

+++ unet/train_exp3.py
@@ -204,16 +204,16 @@ def train():
     print('Loaded train images: ', imgs_train.shape, imgs_mask_train.shape)
     print('-'*30)
 
-    imgs_train = imgs_train.astype('float32')
     mean = np.mean(imgs_train)  # mean for data centering
     std = np.std(imgs_train)  # std for data normalization
 
     # Normalization of the train set (Exp 3)
+    imgs_train = imgs_train.astype('float32')
     imgs_train -= mean
     imgs_train /= std
 
-    imgs_train = imgs_train.astype('int16')
-    imgs_mask_train = imgs_mask_train.astype('int16')
+    # imgs_train = imgs_train.astype(np.float32)
+    imgs_mask_train = imgs_mask_train.astype(np.float32)
 
     print('Creating and compiling model...')
     print('-'*30)
